examples: update react graph example, use nonworker version
