+++ examples/react-graph/package.json
@@ -14,7 +14,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@pattaya/depict": "^0.5.3",
+    "@pattaya/depict": "^0.6.2",
     "impressionist": "^1.0.4",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
