+++ examples/react-graph/src/graph/bar.ts
@@ -1,5 +1,5 @@
-import { ShadowElement } from "@pattaya/depict/graph";
-import { reactState } from "./state";
+import type { ShadowElement } from "@pattaya/depict/graph";
+import { graphState } from "./state";
 
 export const barGraph: ShadowElement[] = [
   {
@@ -19,7 +19,7 @@ export const barGraph: ShadowElement[] = [
     ],
     data: { rotation: 0, timestamp: 0 },
     update(timestamp: number) {
-      this.data.rotation += (reactState.count % 10 + 3) * (timestamp - this.data.timestamp) / 3000;
+      this.data.rotation += (graphState.count % 10 + 3) * (timestamp - this.data.timestamp) / 3000;
       this.shapes![0].opts!.rotation = this.data.rotation;
       this.data.timestamp = timestamp;
     }
