+++ examples/react-graph/src/graph/state.ts
@@ -1,0 +1,3 @@
+export const graphState = {
+  count: 0,
+};
