--- examples/react-graph/src/layers/state.ts
@@ -1,3 +1,0 @@
-export const reactState = {
-  count: 0,
-};
